requirements.txt
google-generativeai>=0.3.1
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
